// src/enums/ColumnsReqTypes.ts
export enum ColumnsReqTypes {
    TitleEn = 'Title_en',
    FileTemplate = 'FileTemplate',
    FileTemplateEn = 'FileTemplateEn',
    Description = 'Description',
    DescriptionEn = 'Description_en',
    Stage = 'Etapa',
    Fields = 'Fields',
    FileAttachments = 'FileAttachments',
    Notification = 'Notification',
    FileReply = 'FileReply'
}